+++ components/settings/projects/components/CreateProjectForm.tsx
@@ -1,11 +1,11 @@
 import { Box, Stack } from '@mui/material';
 import { useFormContext } from 'react-hook-form';
 
-import { useCreateProject, useGetProjects } from 'charmClient/hooks/projects';
+import { useCreateProject } from 'charmClient/hooks/projects';
 import { Button } from 'components/common/Button';
 import type { ProjectWithMembers, ProjectAndMembersPayload } from 'lib/projects/interfaces';
 
-import { SettingsProjectFormAnswers } from './ProjectForm';
+import { ProjectFormAnswers } from './ProjectForm';
 
 export function CreateProjectForm({
   onCancel,
@@ -28,7 +28,7 @@ export function CreateProjectForm({
 
   return (
     <>
-      <SettingsProjectFormAnswers isTeamLead />
+      <ProjectFormAnswers isTeamLead />
       <Box
         sx={{
           display: 'flex',
